Bug 497556 - LDAPI connections cause TCP performance degradation
If a slot from the connection table had previously been used for
an LDAPI connection, it would cause serious performance degradation
(20 times worse on my system) to a TCP connection that reused the
same slot.  The problem was that we were not clearing the flag
that indicates that a connection is a UNIX local socket in the
connection cleanup function.  Clearing this flag makes the reused
slots perform as expected.