Set unknown attribute syntax to Octet String.
When an unknown attribute type is used (when adding an entry using
the extensibleObject objectclass for example), the syntax of the
attribute defaults to "Directory String".  This places syntax
restrictions on the attribute, limiting the attribtues use to
specific types of values.  This patch changes the syntax for
unknown attribute to "Octet String", which allows 0 or more UTF-8
bytes.  This will allow unknown attributes to store binary or empty
values.