Bug 536703 - Don't send empty mod to AD for mapped DN values
When using winsync, setting a mapped DN attribute (such as seeAlso) to
a DN outside the scope of the sync agreement causes an empty modify
operation to be sent to AD.  This causes AD to respond with LDAP error
89, which triggers the update to abort.

The problem is that windows_update_remote_entry() uses a flag to
determine if it has a modify to send to AD.  This flag is set by
windows_generate_update_mods(), but the mods are empty since it is
detecting that the DN used in seeAlso is not in the scope of the sync
agreement.

The fix is to only set the modify flag if we actually have mods to send.