Bug 537466 -  nsslapd-distribution-plugin should not require plugin name to begin with "lib"
https://bugzilla.redhat.com/show_bug.cgi?id=537466
Resolves: bug 537466
Bug Description: nsslapd-distribution-plugin should not require plugin name to begin with "lib"
Reviewed by: nhosoi (Thanks!)
Branch: HEAD
Fix Description: We use the function PL_GetLibraryName from NSPR to get the
platform specific DLL/shared lib filename suffix.  Unfortunately, this function
also prepends the string "lib" to the given name if the given file has no
suffix.  If the given name already has the correct suffix, it does not prepend
the "lib" to the name.  get_plugin_name() should look for the original
library name in the string returned by PL_GetLibraryName.  If it is there, and
has something before it in the string, see if it is "/lib" - if so, remove the
extraneous "lib" string.  If "/lib" is not there, then just pass the string
through as is.
To summarize:
/full/path/to/name.so -> /full/path/to/name.so - error if /full/path/to/name.so does not exist
name -> /default/plugin/path/libname.so -> /default/plugin/path/name.so
name.so -> /default/plugin/path/name.so
Platforms tested: RHEL5 x86_64
Flag Day: no
Doc impact: no