Bug 549554 - Trim single-valued attributes before sending to AD
There are a number of attributes that AD defines as single-valued
that are multi-valued in the LDAP RFCs.  We already had a few
special cases in the winsync code where we only send one attribute
to AD to allow the change to be accepted.  We then simply check if
the value in AD is present in DS when comparing entries for further
changes.

This fix cleans up that old implementation a bit by adding a static
list of these single-valued attributes and a helper function to
check if a particular type is in that list.  I also had to extend
the attr_compare_present() function to allow a length to be passed
in for limiting the comparison to a portion of the values.  This is
needed for the initials attribute, which is single-valued and length
constrained in AD.