fix memory leak in attr replace when replacement fails
if replacement of the attribute values fails (e.g. due to duplicate values)
the valstoreplace is not freed - the caller expects the valueset_replace
function to own the values passed in.  The function will now free the values
if there was an error
In addition, valueset_replace should not free the old values in case
of error - it should leave the old values in the attribute
Reviewed by: nhosoi (Thanks!)
(cherry picked from commit 43894ffddf76baa4824c93d702ad4712e82b5b4e)