Bug 570912 - Avoid selinux context conflict with httpd
One of the dirsrv selinux module interfaces used by the admin
server creates a conflict with the httpd policy.  This change
pulls out the conflicting rule from the interface used to extend
the httpd policy.  A new interface is available with the rule that
was pulled out for use by the admin server CGIs (which causes no
conflict for httpd).