Allow instance name to be parsed from start-slapd
The admin server CGIs need to be able to easily parse the
instance name from the start-slapd script.  Recent format
changes have caused the existing parsing to break, so this
patch makes the parsing of the instance name easier.

To deal with the change in start-slapd format for an upgraded
instance, I have changed the setup code to regenerate all of
the instance scripts during an upgrade instead of simply adding
missing scripts.  This is needed for any bug fix that modifies
a script template to work for an upgraded instance.  I also
added code to write the instance sysconfig script during upgrade
if it doesn't exist already.  We don't want to overwrite this
file if it already exists since it's designed for local changes
to be made to it.