Add managed entries plug-in
This adds a new managed entries plug-in.  This plug-in allows
one to have the Directory Server automatically maintain a set
of entries that are based off of another type of entry (such
as user private group entries based off of user entries).

For more details, see the design document at:

  http://directory.fedoraproject.org/wiki/Managed_Entry_Design