Bug 578863 - Password modify extop needs to send referrals on replicas
The password modify extended operation was modifying the local database
on a read-only replica instead of returning a referral.  The server is
designed to let the plugin ID used for updating password retry info make
local updates instead of returning a referral.  This plugin ID was being
used by the password extop code, which it should not be doing.

The second issue is that we need to check if a referral needs to be sent
as early as possible when processing the extop request.  We don't want
to reject the change if an entry does not exist before checking if a
referral is necessary since the server we refer to may have the target
entry present.  This required adding a new helper function that allows
one to see if a write operation to a particular DN would require a
referral to be sent.  The password modify extop code leverages this new
function to get the referrals and return them to the client if necessary.