Bug 584156 - Remove ldapi socket file during upgrade
The ldapi socket file is only removed when ns-slapd is started
since the server does not have permission to remove it at shutdown.
The causes issues when upgrading to a recetn version that has
SELinux policy since the newly confined ns-slapd daemon will not
be allowed to remove the old ldapi socket file since it doesn't
have a dirsrv specific label.  To deal with this, I've added an
upgrade scriptlet that will remove the ldapi socket file.  When
the newly confined ns-slapd starts up, it will create a new
socket file with the proper label.