Fix rsearch usage of name files for random filters
The rsearch utility has an option (-i) that allows one to have a
file containing names to be used as a variable in search filters.
The way this is designed is that you would specify a partial filter
(such as "uid=") as the filter option and a random name from the
name file will be appended to the filter for each search operation.

Unfortunately, this option was broken.  The format was missing
where we were creating the filter string, which caused the filter
option to be used as specified without appending a name from the
name file.