585905 - ACL with targattrfilters error crashes the server
https://bugzilla.redhat.com/show_bug.cgi?id=585905

Bug Description:
targattrfilters takes this format of value:
 (targattrfilters="add=attr1:F1 && attr2:F2... &&
  attrn:Fn,del=attr1:F1 && attr2:F2 ... && attrn:Fn")
The ACL plugin code had blindly expected the value contains
the operator "add" or "del" and '=' to concatenate the
attribute and filter pair.  The plugin should have checked
the possibility that the value does not follow the format.

Fix Description:
If '=' is not included in the targattrfilters value, the
ACL parser returns ACL_SYNTAX_ERR.  Also, adding a check
code for the returned pointer from strchr and strstr.