574167 - An escaped space at the end of the RDN value is not handled correctly
https://bugzilla.redhat.com/show_bug.cgi?id=574167

Bug Description: If a DN contains "\ " at the end of its RDN,
it's be converted to "\20" by slapi_dn_normalize_ext in the
add operation.  But the following search returns ' ' (not
an escaped space).

Fix Description: When slapi_dn_normalize_ext was applied to a
string which contains "\20", it converted the string to ' '.
This fix changes the behaviour so that the string "\20" in
DN remains untouched.

Also, this patch includes a fix to add a default suffix value
with no double quotes in template-suffix-db.ldif.in.  We keep
double quoted suffix for the backward compatibility.