Bug 584497 - Allow DNA plugin to set same value on multiple attributes
This patch allows one to set multiple dnaType attributes for a single
DNA range.  This allows the same value to be used for each dnaType
(such as ensuring the uidNumber and gidNumber are the same value
from the range when ading a posixAccount).

There are some differences with the way DNA works for multi-type
ranges.  For a value to be generated from a multi-type range, the
magic value must be specified to indicate which attributes should
use the newly generated value.  This allows a range to be shared
across different entry types (such as having a uidNumber/gidNumber
range that is used or posixAccount and posixGroup entries).  A
multi-type range will not generate a value for missing attributes
as DNA does for single-type ranges.

Since a range can have multiple types, the internal ordering of
config structs had to be changed to order by scope only.  It was
previously ordered by type and then scope, which allowed us to
easily detect if we had already generated a value for a given type.
This change required a number of helper functions to be added for
tracking and checking which types we have already generated values
for.