593453 - Creating password policy with ns-newpolicy.pl on Replicated          Server Causes Crash
Fix description: normalized_mods->mod_bvalues is an array of
(struct berval *).  When slapi_dn_normalize_ext returns an
allocated memory (case rc > 0), a new struct berval should have
been allocated and the address of the normalized string (normed)
and its length (dnlen) should be set in the berval.
The struct berval allocation was missing in the previous fix
for "Update to New DN Format".