610281 - fix coverity Defect Type: Control flow issues
https://bugzilla.redhat.com/show_bug.cgi?id=610281

11816 DEADCODE Triaged Unassigned Bug Moderate Fix Required
NS7bitAttr_Init() ds/ldap/servers/plugins/uiduniq/7bit.c

Comment:
NS7bitAttr_Init declared err twice.  One at the top and another
in BEGIN - END (do - while loop).  The second err in the do -
while loop is trashed when it gets out of the loop.  Regardless
of the result in the do - while loop, err = 0 (success) was
returned to the caller.

We are removing the second err in the BEGIN (== do - while) scope.
692 int err;