610281 - fix coverity Defect Type: Control flow issues
https://bugzilla.redhat.com/show_bug.cgi?id=610281

11829 DEADCODE Triaged Unassigned Bug Minor Fix Required
ldbm_back_start() ds/ldap/servers/slapd/back-ldbm/start.c

Comment:
The code meant the autosized cache should be reduced 25%
if the calculated size was less than 500MB for the overhead
of libdb.  If larger than 500MB, the overhead is relatively
small and can be ignored.  The code should have calculated
the size before comparing it to 500MB.