Bug 612242 - membership change on DS does not show on AD
When a change was made to a DN mapped attribute in DS (such as
uniqueMember in a group entry), we may end up searching for the
entries that those values point to in AD when winsync is being
used.  We were overwriting the "raw entry" pointer every time we
searched for an entry in AD.  The raw entry is intended to point
to the entry that the original modification was made to, not the
entry that a DN mapped attribute value points to.

The fix is to add a flag that will force the raw entry to be kept
when we search for an entry in AD.  We set this flag when we search
for entries that are pointed to be DN mapped attribute values and
reset it when we are finished.  This results in the raw entry being
the actual entry that is the target of the operation we are syncing.