Bug 547503  - replication broken again, with 389 MMR replication and TCP errors
https://bugzilla.redhat.com/show_bug.cgi?id=547503
Resolves: bug 547503
Bug Description: replication broken again, with 389 MMR replication and TCP errors
Reviewed by: nhosoi (Thanks!)
Branch: 389-ds-base-1.2.6
Fix Description: When turbo mode is used for the connection, the server does
not poll for read ready status in the main loop, nor go through the code in
handle_pr_read_ready that updates conn->c_idlesince.  So while the conn is in
turbo mode, the c_idlesince is not updated.  If the conn gets a timeout while
reading, a flag will be set on the connection that will put it back in the
main loop.  When it then hits handle_pr_read_ready, if there is still no
activity on the connection, it will go through idle timeout processing.  It
may have been a long time since c_idlesince was updated, so the connection
may be closed wrongly.
The solution is to have c_idlesince updated in connection_threadmain() in
turbo mode if the connection really isn't idle.
In addition, the conn private turbo_mode flag was not being used correctly -
in the timeout case, the local variable was being updated but not the
conn private turbo_flag.  Since the conn private turbo_flag is not used
anywhere else, it can be removed, and just use the local variable.
Platforms tested: RHEL5 x86_64
Flag Day: no
Doc impact: no