Bug 617630 - fix coverify Defect Type: Resource leaks issues CID 12052 - 12093
https://bugzilla.redhat.com/show_bug.cgi?id=617630
Resolves: bug 617630
Bug description: fix coverify Defect Type: Resource leaks issues CID 12086, 12091, 12092.
description: The vlv_build_idl() has been modified to release idl when an error occurs. The "if (idl)" near the end of the function is unnecessary because at that point idl cannot be NULL. The "if (candidate)" has been inverted to catch NULL pointer.

The vlv_filter_candidates() and the vlv_trim_candidates() have been modified to check for NULL filteredCandidates in the beginning to avoid resource leaks.